+++ R/plotScores.R
@@ -33,7 +33,7 @@
 #'
 #' @param tol A number describing the fraction of points to be labeled.
 #' \code{tol = 1.0} labels all the points; \code{tol = 0.05} labels the most
-#' extreme 5 percent.
+#' extreme 5 percent.  Set to \code{"none"} to completely suppress labels.
 #'
 #' @param use.sym A logical; if TRUE, the color scheme is set to black and the
 #' points plotted with symbols.  Applies only to \code{\link[ChemoSpec]{Spectra}} objects.